黃威誠.新增: SpringMVC換到SpringBoot

+++ src/main/java/com/booking/utils/Listable.java
@@ -6,6 +6,6 @@ import java.util.Map;
 public interface Listable {
 	String getName();
 	Integer getId();
-	Integer getTotalCounts();
+	Long getTotalCounts();
 	Map<String, Object> getAdditionProperties();
 }
